New test and yaml for verifiable dataset
