+++ .github/workflows/pq_pipeline.yml
@@ -105,6 +105,16 @@ jobs:
     with:
       commit_id: ${{ needs.set_commit_id_for_all_jobs.outputs.commit_id }}
 
+  tr_verifiable_dataset_e2e:
+    if: |
+      (github.event_name == 'schedule' && github.repository_owner == 'securefederatedai') ||
+      (github.event_name == 'workflow_dispatch')
+    name: TaskRunner Verifiable Dataset E2E
+    needs: task_runner_e2e
+    uses: ./.github/workflows/tr_verifiable_dataset_e2e.yml
+    with:
+      commit_id: ${{ needs.set_commit_id_for_all_jobs.outputs.commit_id }}
+
   task_runner_secure_agg_e2e:
     if: |
       (github.event_name == 'schedule' && github.repository_owner == 'securefederatedai') ||
